requirements.txt
flask==3.0.3
pillow==10.4.0
pymupdf==1.24.9
